Refactor type and role retrieval in diagnosis example
